added purchase,borrowing and wallet features

+++ routes/api.php
@@ -2,10 +2,9 @@
 
 use App\Http\Controllers\AuthorController;
 use App\Http\Controllers\BookController;
+use App\Http\Controllers\BookCopyController;
 use App\Http\Controllers\OwnerController;
 use App\Http\Controllers\UserController;
-use App\Http\Middleware\EnsureIsOwner;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\BookRatingController;
 use App\Http\Controllers\CategoryController;
@@ -13,44 +12,76 @@ use App\Http\Controllers\OrderController;
 use App\Http\Controllers\FavoriteController;
 use App\Http\Controllers\MyReadingController;
 use App\Http\Controllers\VisitReservationController;
-
+use App\Http\Controllers\WalletController;
+use App\Http\Controllers\BorrowReservationController;
+use App\Http\Controllers\BorrowController;
+use App\Http\Controllers\PurchaseController;
+use App\Http\Controllers\DigitalProductsController;
+use App\Http\Controllers\ProfileController;
+use App\Http\Controllers\ContactController;
 Route::post('owner/login',[OwnerController::class,'login']);
 
 Route::prefix('owner')->middleware(['auth:sanctum','ensure.owner'])->group(function (){
      Route::post('/visits', [VisitReservationController::class, 'store']);
      Route::get('/visits', [VisitReservationController::class, 'index']);
+     Route::get('/visits/{id}', [VisitReservationController::class, 'show']);
+Route::post('/visits/check', [VisitReservationController::class, 'check']);
 Route::put('/visits/{id}/status', [VisitReservationController::class, 'updateStatus']);
+Route::get('/borrowReservations',[BorrowReservationController::class,'index']);
+Route::get('/borrowReservations/{id}',[BorrowReservationController::class,'show']);
+Route::post('/borrowReservations/{id}', [BorrowController::class, 'reservationToBorrowing']);
+Route::post('/walkInBorrowing', [BorrowController::class, 'walkInBorrowing']);
+Route::get('/borrowings', [BorrowController::class, 'index']);
+Route::post('/borrowings/return', [BorrowController::class, 'returnBorrowing']);
+Route::post('/walkInPurchases',[PurchaseController::class,'walkInPurchase']);
+Route::get('/purchases',[PurchaseController::class,'index']);
 
 
+    Route::post('/categories', [CategoryController::class, 'store']);       // POST /api/categories
+    Route::put('/categories/{id}', [CategoryController::class, 'update']);   // PUT /api/categories/{id}
+    Route::delete('/categories/{id}', [CategoryController::class, 'destroy']);// DELETE /api/categories/{id}
+    Route::post('/books', [BookController::class, 'store']);
+    Route::put('/books/{id}', [BookController::class, 'update']);
+    Route::delete('/books/{id}', [BookController::class, 'destroy']);
+    Route::post('/authors', [AuthorController::class, 'store']);       // POST /api/authors
+    Route::put('/authors/{id}', [AuthorController::class, 'update']);   // PUT /api/authors/{id}
+    Route::delete('/authors/{id}', [AuthorController::class, 'destroy']);// DELETE /api/authors/{id}
+
+    Route::prefix('wallets')->group(function () {
+    Route::get('search', [WalletController::class, 'search']);
+    Route::get('{wallet}', [WalletController::class, 'show']);
+    Route::post('{wallet}/charge', [WalletController::class, 'charge']);
+    Route::post('{wallet}/freeze', [WalletController::class, 'freeze']);
+    Route::post('{wallet}/unfreeze', [WalletController::class, 'unfreeze']);
+    Route::post('{wallet}/buy', [WalletController::class, 'buyWithCredits']);
+
+});
+
+Route::post('/checkPassword',[OwnerController::class,'checkPassword']);
+
+Route::post('/getBookByCopyBarcode',[BookCopyController::class,'getBookByCopyBarcode']);
 });
 
 
 //books routs
 Route::prefix('books')->group(function () {
     Route::get('/', [BookController::class, 'index']);
-    Route::post('/', [BookController::class, 'store']);
     Route::get('/latest', [BookController::class, 'latestReleases']);
     Route::get('/bestsellers', [BookController::class, 'bestSellers']);
     Route::get('/search', [BookController::class, 'search']); //Search route
     Route::get('/{id}', [BookController::class, 'show']);
-    Route::put('/{id}', [BookController::class, 'update']);
-    Route::delete('/{id}', [BookController::class, 'destroy']);
 });
 //authors routes
 Route::prefix('authors')->group(function () {
     Route::get('/', [AuthorController::class, 'index']);        // GET /api/authors
-    Route::post('/', [AuthorController::class, 'store']);       // POST /api/authors
     Route::get('/{id}', [AuthorController::class, 'show']);     // GET /api/authors/{id}
-    Route::put('/{id}', [AuthorController::class, 'update']);   // PUT /api/authors/{id}
-    Route::delete('/{id}', [AuthorController::class, 'destroy']);// DELETE /api/authors/{id}
+
 });
 //categories routes
 Route::prefix('categories')->group(function () {
     Route::get('/', [CategoryController::class, 'index']);        // GET /api/categories
-    Route::post('/', [CategoryController::class, 'store']);       // POST /api/categories
     Route::get('/{id}', [CategoryController::class, 'show']);     // GET /api/categories/{id}
-    Route::put('/{id}', [CategoryController::class, 'update']);   // PUT /api/categories/{id}
-    Route::delete('/{id}', [CategoryController::class, 'destroy']);// DELETE /api/categories/{id}
+
 });
 //user authentication routes
 Route::controller(UserController::class)->group(function(){
@@ -103,3 +134,32 @@ Route::middleware('auth:sanctum')->prefix('myreading')->group(function () {
     Route::delete('/{id}', [MyReadingController::class, 'destroy']);
 });
 
+// visits routes
+Route::middleware(['auth:sanctum'])->group(function () {
+    Route::post('/visits', [VisitReservationController::class, 'store']);
+    Route::delete('/visits/{id}', [VisitReservationController::class, 'cancel']);
+         Route::get('/visits/{id}', [VisitReservationController::class, 'show']);
+    Route::post('/visits/checkAvailableTimes', [VisitReservationController::class, 'checkAvailableTimes']);
+    Route::get('/visits/myReservations',[VisitReservationController::class,'myReservations']);
+});
+
+//user wallet
+Route::middleware(['auth:sanctum'])->group(function () {
+    Route::get('/wallet',[WalletController::class,'showUserWallet']);
+});
+//digital products
+Route::middleware(['auth:sanctum'])->prefix('digitalProducts')->group(function () {
+    Route::get('/', [DigitalProductsController::class, 'index']);        // GET /api/product
+    Route::post('/', [DigitalProductsController::class, 'store']);       // POST /api/product
+    Route::get('/{id}', [DigitalProductsController::class, 'show']);     // GET /api/product/{id}
+    Route::put('/{id}', [DigitalProductsController::class, 'update']);   // PUT /api/product/{id}
+    Route::delete('/{id}', [DigitalProductsController::class, 'destroy']);// DELETE /api/product/{id}
+});
+//user profile
+Route::middleware(['auth:sanctum'])->prefix('profile')->group(function () {
+    Route::get('/{id}', [ProfileController::class, 'show']);     // GET /api/Profile /{id}
+    Route::put('/{id}', [ProfileController::class, 'update']);   // PUT /api/Profile/{id}
+    Route::delete('/{id}', [ProfileController::class, 'destroy']);// DELETE /api/Profile/{id}
+});
+
+Route::middleware(['auth:sanctum'])->post('/contact', [ContactController::class, 'send']);
